Update Configure.SCVMM.Core.Requirements.ps1
Added '/norestart' for both VisualC2017 .exe files in ArgumentList parament. Already there on the ADK and WinPE blocks.

Added 'REBOOT=ReallySuppress' in both msiexec calls ArgumentList to suppress a premature reboot there. 

Added "Installing ..." lines with Magenta color so User can see what is happening before it starts. 

Modified the Restart-Computer block so the user is prompted to reboot the computer or not, in case they need to do something else first.